requirements.txt
elasticsearch==8.13.0
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.4.2
joblib==1.4.2
python-dotenv==1.0.1
google-generativeai==0.7.2
requests==2.32.3
